+++ src/main/java/com/phantomwing/choppersdelight/datagen/ModBlockTagsProvider.java
@@ -1,10 +1,13 @@
 package com.phantomwing.choppersdelight.datagen;
 
 import com.phantomwing.choppersdelight.ChoppersDelight;
+import com.phantomwing.choppersdelight.Compatibility;
 import com.phantomwing.choppersdelight.block.ModBlocks;
 import com.phantomwing.choppersdelight.tags.ModTags;
+import com.phantomwing.choppersdelight.utils.BlockUtils;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
+import net.minecraft.resources.ResourceLocation;
 import net.neoforged.neoforge.common.data.BlockTagsProvider;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
 import org.jetbrains.annotations.NotNull;
@@ -23,9 +26,28 @@ public class ModBlockTagsProvider extends BlockTagsProvider {
             .add(vectorwing.farmersdelight.common.registry.ModBlocks.CUTTING_BOARD.get())
             .add(ModBlocks.DECORATED_CUTTING_BOARD.get());
 
-        ModBlocks.CUTTING_BOARDS.forEach((blockSupplier -> {
-            this.tag(ModTags.Blocks.CUTTING_BOARDS).add(
-                    blockSupplier.get()
+        // Add Vanilla cutting boards
+        ModBlocks.MINECRAFT_CUTTING_BOARDS.forEach((blockSupplier -> {
+            this.tag(ModTags.Blocks.CUTTING_BOARDS).add(blockSupplier.get());
+        }));
+
+        // Add Biomes O' Plenty cutting boards
+        ModBlocks.BIOMES_O_PLENTY_CUTTING_BOARDS.forEach((blockSupplier -> {
+            this.tag(ModTags.Blocks.CUTTING_BOARDS).addOptional(
+                ResourceLocation.fromNamespaceAndPath(
+                    Compatibility.BIOMES_O_PLENTY_MOD_ID,
+                    BlockUtils.getName(blockSupplier.get())
+                )
+            );
+        }));
+
+        // Add Biomes We've Gone cutting boards
+        ModBlocks.BIOMES_WEVE_GONE_CUTTING_BOARDS.forEach((blockSupplier -> {
+            this.tag(ModTags.Blocks.CUTTING_BOARDS).addOptional(
+                ResourceLocation.fromNamespaceAndPath(
+                    Compatibility.BIOMES_WEVE_GONE_MOD_ID,
+                    BlockUtils.getName(blockSupplier.get())
+                )
             );
         }));
     }