+++ src/main/java/com/phantomwing/choppersdelight/integration/JEIPlugin.java
@@ -1,6 +1,6 @@
 package com.phantomwing.choppersdelight.integration;
 
-import com.phantomwing.choppersdelight.block.ModBlocks;
+import com.phantomwing.choppersdelight.utils.BlockUtils;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
 import mezz.jei.api.registration.*;
@@ -27,9 +27,8 @@ public class JEIPlugin implements IModPlugin
 
     @Override
     public void registerRecipeCatalysts(IRecipeCatalystRegistration registration) {
-        ModBlocks.CUTTING_BOARDS.forEach((blockSupplier) -> {
-            registration.addRecipeCatalyst(new ItemStack(blockSupplier.get()), FDRecipeTypes.CUTTING);
+        BlockUtils.getCuttingBoards().forEach((block) -> {
+            registration.addRecipeCatalyst(new ItemStack(block), FDRecipeTypes.CUTTING);
         });
-
     }
 }